requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
django-htmx==1.17.2